Revert "Move credential schemas and modules into centralized datastore module"
This reverts commit 311d5158c0e09839b8c259da4d6a6f76bc420e22.
