+++ libsplinter/src/biome/credentials/store/diesel/operations/get_usernames.rs
@@ -12,10 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::super::{CredentialsStoreError, UsernameId};
+use super::super::super::UserCredentialsModel;
+use super::super::{schema::user_credentials, CredentialsStoreError, UsernameId};
 use super::CredentialsStoreOperations;
-use crate::biome::datastore::models::UserCredentialsModel;
-use crate::biome::datastore::schema::user_credentials;
 use diesel::prelude::*;
 
 pub(in super::super) trait CredentialsStoreGetUsernamesOperation {
