+++ libsplinter/src/biome/credentials/store/diesel/operations/remove_credentials.rs
@@ -12,10 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::super::CredentialsStoreError;
+use super::super::super::UserCredentialsModel;
+use super::super::{schema::user_credentials, CredentialsStoreError};
 use super::CredentialsStoreOperations;
-use crate::biome::datastore::models::UserCredentialsModel;
-use crate::biome::datastore::schema::user_credentials;
 use diesel::{dsl::delete, prelude::*, result::Error::NotFound};
 
 pub(in super::super) trait CredentialsStoreRemoveCredentialsOperation {
