+++ libsplinter/src/biome/datastore/diesel/mod.rs
@@ -12,6 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod models;
 pub mod postgres;
-pub mod schema;
